+++ R_package/src/bitoperation.h
@@ -1,10 +1,8 @@
 #pragma once
 
-#include "Int.h"
-
 namespace BitOperation {
 	// ex) 3 -> 1...111000
-	inline Int::ull upper_mask(std::size_t k) {
+	inline uint64_t upper_mask(std::size_t k) {
 		if(k == 0)
 			return ~0;
 		else
@@ -12,12 +10,12 @@ namespace BitOperation {
 	}
 	
 	// ex) 3 -> 0...000111
-	inline Int::ull lower_mask(std::size_t k) {
+	inline uint64_t lower_mask(std::size_t k) {
 		return (1ULL << k) - 1;
 	}
 	
 	// ex) 3, 5 -> 0...00011000
-	inline Int::ull middle_mask(std::size_t first, std::size_t last) {
+	inline uint64_t middle_mask(std::size_t first, std::size_t last) {
 		return ((1ULL << (last - first)) - 1) << first;
 	}
 }
